requirements.txt
# QwenFlow Backend Dependencies
# Python 3.9+ required

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0

